X-Lovable-Edit-ID: edt-b3abaf80-56fd-457a-9099-49877ac7414e

+++ src/routes/route.$token.tsx
@@ -191,7 +191,7 @@ function FollowerPage() {
     !expiresAt && createdAt
       ? now.getTime() - createdAt.getTime() > 30 * 24 * 60 * 60 * 1000
       : false;
-  const dateFmt = (d: Date) => d.toLocaleDateString(undefined, { day: "numeric", month: "short", year: "numeric" });
+  const dateFmt = (d: Date) => d.toLocaleDateString(undefined, { day: "numeric", month: "long", year: "numeric" });
 
   return (
     <div className="h-screen flex flex-col overflow-hidden">
